+++ nexus-core/nexus-core-security/src/main/java/com/github/arpit/nexus/core/security/ratelimit/RateLimit.java
@@ -1,0 +1,15 @@
+package com.github.arpit.nexus.core.security.ratelimit;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import java.util.concurrent.TimeUnit;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RateLimit {
+    int limit() default 10;
+    int duration() default 60;
+    TimeUnit unit() default TimeUnit.SECONDS;
+}